requirements.txt


# fastapi==0.115.6
# uvicorn==0.34.0
# dvc==2.8.3

# python 3.12.8

# pytorch==2.2.2


numpy<2
typer==0.15.1
torch==2.2.2
pandas==2.2.3
transformers==4.48.0
tf-keras==2.18.0
fsspec[http]>=2023.1.0,<2024.6.1
accelerate>=0.26.0
hydra-core==1.3.2
torch_tb_profiler==0.4.3

# for cloud storage
dvc[gdrive]==3.59.0
dvc_gs==3.0.1

# for code logging
loguru==0.7.2

# sentencepiece
sentencepiece==0.2.0

# for testing code
pytest==8.3.4

#Code Coverage
coverage==7.6.10